Refactor: move media models to data.models
