+++ app/src/main/java/fr/free/nrw/commons/data/models/media/IdAndCaptions.kt
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.media
+package fr.free.nrw.commons.data.models.media
 
 data class IdAndCaptions(val id: String, val captions: Map<String, String>)
