+++ app/src/main/java/fr/free/nrw/commons/media/MediaDetailFragment.java
@@ -10,10 +10,9 @@ import static fr.free.nrw.commons.description.EditDescriptionConstants.LIST_OF_D
 import static fr.free.nrw.commons.description.EditDescriptionConstants.UPDATED_WIKITEXT;
 import static fr.free.nrw.commons.description.EditDescriptionConstants.WIKITEXT;
 import static fr.free.nrw.commons.upload.mediaDetails.UploadMediaDetailFragment.LAST_LOCATION;
-import android.content.res.Resources;
 import static fr.free.nrw.commons.utils.LangCodeUtils.getLocalizedResources;
 import android.annotation.SuppressLint;
-import java.lang.reflect.Field;
+import fr.free.nrw.commons.data.models.media.Caption;
 import android.app.AlertDialog;
 import android.content.Context;
 import android.content.Intent;
@@ -75,6 +74,7 @@ import fr.free.nrw.commons.category.CategoryEditSearchRecyclerViewAdapter;
 import fr.free.nrw.commons.category.CategoryEditSearchRecyclerViewAdapter.Callback;
 import fr.free.nrw.commons.contributions.ContributionsFragment;
 import fr.free.nrw.commons.coordinates.CoordinateEditHelper;
+import fr.free.nrw.commons.data.models.media.IdAndCaptions;
 import fr.free.nrw.commons.delete.DeleteHelper;
 import fr.free.nrw.commons.delete.ReasonBuilder;
 import fr.free.nrw.commons.description.DescriptionEditActivity;
@@ -93,7 +93,6 @@ import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.schedulers.Schedulers;
 import java.util.ArrayList;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Locale;
